+++ workflow/tests/util.rb
@@ -100,14 +100,14 @@ def _run_xml(xml, worker_num, apply_unit_multiplier = false, annual_results_1x =
   # Check outputs
   hpxml_defaults_path = File.join(rundir, 'in.xml')
   schema_validator = XMLValidator.get_xml_validator(File.join(File.dirname(__FILE__), '..', '..', 'HPXMLtoOpenStudio', 'resources', 'hpxml_schema', 'HPXML.xsd'))
-  schematron_validator = XMLValidator.get_xml_validator(File.join(File.dirname(__FILE__), '..', '..', 'HPXMLtoOpenStudio', 'resources', 'hpxml_schematron', 'EPvalidator.xml'))
+  schematron_validator = XMLValidator.get_xml_validator(File.join(File.dirname(__FILE__), '..', '..', 'HPXMLtoOpenStudio', 'resources', 'hpxml_schematron', 'EPvalidator.sch'))
   hpxml = HPXML.new(hpxml_path: hpxml_defaults_path, schema_validator: schema_validator, schematron_validator: schematron_validator) # Validate in.xml to ensure it can be run back through OS-HPXML
   if not hpxml.errors.empty?
     puts 'ERRORS:'
     hpxml.errors.each do |error|
       puts error
     end
-    flunk "EPvalidator.xml error in #{hpxml_defaults_path}."
+    flunk "EPvalidator.sch error in #{hpxml_defaults_path}."
   end
   annual_results = _get_simulation_annual_results(annual_csv_path, bills_csv_path)
   monthly_results = _get_simulation_monthly_results(monthly_csv_path)